refactor: use JS modules on personal diary page
