+++ diaries/christian.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="/src/styles/global.css" />
     <link rel="stylesheet" href="/src/styles/diaries.css" />
     <link rel="stylesheet" href="http://fonts.googleapis.com/css?family=Bree+Serif" />
-    <script defer src="/src/scripts/diary.js"></script>
+    <script type="module" defer src="/src/scripts/diary.js"></script>
   </head>
 
   <body data-person="Christian">
